docs: add oversized label demo
